requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
chromedriver-autoinstaller==0.4.0
Django==4.1.2
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
selenium==4.4.3
openpyxl==3.0.10
